# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger for Fiber",
        "title": "Fiber Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "fiber@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/api",
    "paths": {
...
        "/card/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all cards from auth user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "card"
                ],
                "summary": "Get auth user cards",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.CardsRespOKDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.AuthRespErrDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.DetailRespErrDTO"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Save card to user acount",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "card"
                ],
                "summary": "Save card",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.CardsRespOKDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.AuthRespErrDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.DetailRespErrDTO"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove a card from user acount",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "card"
                ],
                "summary": "Remove card",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.CardsRespOKDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.AuthRespErrDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.DetailRespErrDTO"
                        }
                    }
                }
            }
        },
...
        "/order/new": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "Create new order",
                "parameters": [
                    {
                        "description": "order data",
                        "name": "order_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.NewOrderDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.OrderDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.ValidationRespErrDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.AuthRespErrDTO"
                        }
                    },
                    "406": {
                        "description": "Not Acceptable",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.RespErrDTO"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.DetailRespErrDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.DetailRespErrDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_ZaphCode_clean-arch_src_api_dtos.AuthRespErrDTO": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "the token is expired"
                },
                "message": {
                    "type": "string",
                    "example": "invalid access token"
                },
                "status": {
                    "type": "string",
                    "example": "failure"
                }
            }
        },
        "github_com_ZaphCode_clean-arch_src_api_dtos.CardDTO": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string",
                    "example": "visa"
                },
                "country": {
                    "type": "string",
                    "example": "USA"
                },
                "created_at": {
                    "type": "integer",
                    "example": 1674405183
                },
                "exp_month": {
                    "type": "integer",
                    "example": 11
                },
                "exp_year": {
                    "type": "integer",
                    "example": 2025
                },
                "id": {
                    "type": "string",
                    "example": "8ded83fe-93c8-11ed-ab0f-d8bbc1a27048"
                },
                "last4": {
                    "type": "string",
                    "example": "1429"
                },
                "name": {
                    "type": "string",
                    "example": "Main card"
                },
                "updated_at": {
                    "type": "integer",
                    "example": 1674405181
                }
            }
        },
        "github_com_ZaphCode_clean-arch_src_api_dtos.CardsRespOKDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.CardDTO"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Data retrived!"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "github_com_ZaphCode_clean-arch_src_api_dtos.DetailRespErrDTO": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "error querying the database"
                },
                "message": {
                    "type": "string",
                    "example": "something went wrong"
                },
                "status": {
                    "type": "string",
                    "example": "failure"
                }
            }
        },
...
        "github_com_ZaphCode_clean-arch_src_api_dtos.NewOrderDTO": {
            "type": "object",
            "required": [
                "address_id",
                "payment_id",
                "products"
            ],
            "properties": {
                "address_id": {
                    "type": "string"
                },
                "payment_id": {
                    "type": "string"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_domain.OrderProduct"
                    }
                }
            }
        },
...
        "github_com_ZaphCode_clean-arch_src_api_dtos.OrderDTO": {
            "type": "object",
            "required": [
                "address_id",
                "payment_id",
                "products"
            ],
            "properties": {
                "address_id": {
                    "type": "string"
                },
                "amount": {
                    "type": "integer",
                    "example": 14500
                },
                "created_at": {
                    "type": "integer",
                    "example": 1674405183
                },
                "id": {
                    "type": "string",
                    "example": "8ded83fe-93c8-11ed-ab0f-d8bbc1a27048"
                },
                "paid": {
                    "type": "boolean",
                    "example": true
                },
                "payment_id": {
                    "type": "string"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_domain.OrderProduct"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "pending"
                },
                "updated_at": {
                    "type": "integer",
                    "example": 1674405181
                }
            }
        },
...
        "github_com_ZaphCode_clean-arch_src_api_dtos.RespErrDTO": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "something went wrong"
                },
                "status": {
                    "type": "string",
                    "example": "failure"
                }
            }
        },
...
        "github_com_ZaphCode_clean-arch_src_api_dtos.ValidationRespErrDTO": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "field": {
                                "type": "string",
                                "example": "Email"
                            },
                            "message": {
                                "type": "string",
                                "example": "invalid email"
                            }
                        }
                    }
                },
                "message": {
                    "type": "string",
                    "example": "one or more field are invalid"
                },
                "status": {
                    "type": "string",
                    "example": "failure"
                }
            }
        },
        "github_com_ZaphCode_clean-arch_src_domain.OrderProduct": {
            "type": "object",
            "properties": {
                "product_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "X-Access-Token",
            "in": "header"
        }
    }
}